cat_backend/server/requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
djangorestframework==3.15.2
future==1.0.0
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3